requirements.txt
Flask==2.1.2
tensorflow
tensorflow-hub
pillow
numpy
python-multipart